logic from custom Hook moved to the component. Custom Hook need to host the vars that will be updated and expose functions to update their states, and in the component only use those states and use the functions to update those states, avoid using another use effect to create new states because that causes asynchronous problems and delays.
